FileManager: Extract a zip archive on double click
Currently, when the user double-clicks a .zip file, it opens TextEditor
as the file format is not associated with any application. Instead, the
user should be able to extract a zip archive from double-click.
As the expected action is less obvious compared to the "Extract Here"
right-click menu, this change also adds a confirmation dialog that's
shown only when the user tries to extract a zip archive from double-
click.
